==> R Scripts RF, MICE, EDA/1. Preliminary Data cleaning with plots.Rmd <==
#data cleaning and plotting for report
#all plots at the end

```{r}
load("~/Thesis R/Unprocessed Data.RData")
rm(longdata)
#load all functions from functions.Rmd
```

#missingness in original data
```{r}
originalmiss = miss_summary(widedata)
sum(originalmiss$`by var`$`Miss no`)/(6640*174) #20.9% of total data missing in the start
originalsector = eda_summary(widedata)

#PLOT 1
barplot(originalmiss$`by firm`$`Miss no`)
barplot(originalmiss$`by var`$`Miss no`)
barplot(originalsector[[3]]$`round((na_bysector[, 2]/na_bysector[, 4]), 2)`)
```
#step 1
#remove anything thats mising in more than half a sequence
```{r}
halfna =  c("Environmental Disclosure Score"
,"ESG Disclosure Score", "Gearing Ratio Percentage Company Basis", "Governance Disclosure Score", "Operating Lease Assets", "Social Disclosure Score", "Total Cash Flow from Op Activities Adjustments", "Total CO2 Emissions", "Total Energy Consumption", "Total Operating Revenue", "Short Term Operating Lease Right Use Assets", "Days Sales Outstanding","Capital Expenditures to Total Assets", "Financing Revenue", "Cost of Financing Revenue", "Short Term Operating Lease Liabilities", "Total Operating Lease Liabilities","Capitalized Interest To Pre-Tax Income", "Capitalized Interest Expense", "Actual Cash Flow per Employee", "Actual Net Income per Employee", "Sales Per 1000 Employees", "Actual Assets per Employee", "Average Age of Assets in Years", "Depreciation Exp To Net Fixed Assets", "Future Minimum Operating Lease Obligations", "Capital Expenditure to Depreciation Expense", "Depreciation Expense To Net Sales", "Depreciation Expenses", "Total Operating Expenses")


wide2 = widedata[,!(names(widedata) %in% halfna)]

miss2 = miss_summary(wide2)
miss2sector = eda_summary(wide2)

sum(miss2$`by var`$`Miss no`)/(6640*144) # 8% misisng only

#PLOT 2
barplot(miss2$`by firm`$`Miss no`, ylim = c(0,3000))
barplot(miss2$`by var`$`Miss no`, ylim = c(0,6000))
barplot(miss2sector[[3]]$`round((na_bysector[, 2]/na_bysector[, 4]), 2)`, ylim = c(0,60))
```

#step 2
```{r}
wideno10 <- subset(wide2, Sector!="10")
wideno105 <- subset(wideno10, Sector!="5")
REincluded= subset(wide2, subset = Ticker =="IRM" | Ticker == "EQIX")
wide3 = rbind(wideno105, REincluded)

rm(REincluded, wideno10, wideno105)

miss3 = miss_summary(wide3)
miss3sector = eda_summary(wide3)

sum(miss3$`by var`$`Miss no`)/(6640*144) # 4% missing only

barplot(miss3$`by firm`$`Miss no`, ylim = c(0,3000))
barplot(miss3$`by var`$`Miss no`, ylim = c(0,6000))
barplot(miss3sector[[3]]$`round((na_bysector[, 2]/na_bysector[, 4]), 2)`, ylim = c(0,60))



plot(sort(miss3$`by firm`$`Miss no`))
```

```{r}
#before removing the next batch
plot(sort(miss3$`by var`$`Miss no`))
abline(h=999)

remove15 = c("Cash Conversion Cycle", "R&D Expense", "Interest Income", "Basic EPS - 5 Year Average Growth", "Incremental Operating Margin", "Accum Depreciation To Gross Fixed Assets", "Accum Depreciation To Total Assets", "Inventory to Cash - Days", "Dividend Coverage Ratio","Inventory Turnover", "Net Repurchases", "After Tax Interest Expense", "EV to Trailing 12M Free Cash Flow to Firm", "T12 FCF to Firm After Operating Lease Payments","Accounts Payable Turnover", "Accounts Payable Turnover Days","Other Operating Income", "Cost of Revenue", "Gross Profit","EPS - 1 Yr Growth", "Number of Treasury Shares")
 
wide4 = wide3[,!(names(wide3) %in% remove15)]

miss4 = miss_summary(wide4)
miss4sector = eda_summary(wide4)

sum(miss4$`by var`$`Miss no`)/(6640*123)

plot(sort(miss4$`by firm`$`Miss no`))
abline(v=127)
```


#wide 5 remove the firms
```{r}
library(tidyverse)
wide5 = filter(wide4, !(Ticker %in% c("AAL", "LYV", "DAL", "VMC","AMD", "FMC", "MO" , "CAH", "EBAY", "SYK", "V", "YUM", "HPQ" )))

miss5 = miss_summary(wide5)
miss5sector = eda_summary(wide5)
sum(miss5$`by var`$`Miss no`)/(5080*124)

save(wide5, file = "Data after removing before imp.RData")
```

#PLOTS are here
#plot 1 shows the comparison between the initial dataset and removing the first 30 variables
```{r}
par(mfrow=c(1,2))
barplot(sort(originalmiss$`by firm`$`Miss no`))
barplot(sort(miss2$`by firm`$`Miss no`), ylim = c(0,3000))

par(mfrow=c(1,2))
barplot(sort(originalmiss$`by var`$`Miss no`), ylim = c(0,6700))
barplot(sort(miss2$`by var`$`Miss no`), ylim = c(0,6700))

par(mfrow=c(1,2))
barplot(sort(originalsector[[3]]$`round((na_bysector[, 2]/na_bysector[, 4]), 2)`))
barplot(sort(miss2sector[[3]]$`round((na_bysector[, 2]/na_bysector[, 4]), 2)`), ylim = c(0,60))
```

#plot for missing variables hajek
```{r}
plot(sort(miss3$`by var`$`Miss%`), type = "o", ylab = "% missing", xlab = "Variable index", main = "Variables sorted by missingness level")
abline(h=15)

```

#plot missing firms
```{r}
plot(sort(miss4$`by firm`$`Miss%`), type = "o", ylab = "% missing", xlab = "Firm index", main = "Firms sorted by missingness level")
abline(h=4.5)
```
#colored removal of variables in one plot
```{r}
x= sort(originalmiss$`by var`$`Miss%`)
cols = c("white", "brown")[(x>45.18)+1]
barplot(x, ylim = c(0,100), col= cols, space = 0, xlab = "Variables", ylab = "% missing", main = "")
legend("topleft", c("After removal", "Before removal"), fill =(col=c("white", "brown")))
```

#colored stack missing by firms
```{r}
y = (rbind((originalmiss$`by firm`$`Miss%`),(miss2$`by firm`$`Miss%`)))
barplot(sort(y[1,]), col = "brown", ylim = c(0,50), space = 0)
par(new= T)
barplot(sort(y[2,]), col = "white", ylim = c(0,50), space = 0, xlab = 'Firms', ylab= "% missing", main = "")
legend("topleft", c("After removal", "Before removal"), fill =(col=c("white", "brown")))

#table showing composition by sector
```
```{r}
(table(train$Sector)/5080)*100
tmp = as.data.frame(cbind(wide5$Ticker, wide5$Sector))


#1 is consumer discretionary
#2 is consumer staples
#3 is communication services
#4 is energy
#5 is financials
#6 is health care
#7 is industrial
#8 is information technology
#9 is materials
#10 is real estate
#11 is utilities

```

```{r}
#load("~/Thesis R/corrected data after imp.RData")
barplot(table(creditr$CR)/50.8, xlab = "Rating Level", ylab = "% of dataset", ylim =c(0,20))
#print(table(creditr$CR)/50.80)
```

```{r}
#load("~/Thesis R/train and test data.RData")
traintest = rbind(train,test)
table(traintest$CR)/50.80
barplot(table(traintest$CR)/50.8, xlab = "Rating Level", ylab = "% of dataset", ylim =c(0,40))
```

```{r}
barplot(table(creditr$CR)/50.8, xlab = "Credit rating level", ylab = "% in dataset", main = "Distribution of credit ratings before grouping")
barplot(table(traintest$CR)/50.8,xlab = "Credit rating level", ylab = "% in dataset", main = "Distribution of credit ratings after grouping")
```
