==> R Scripts RF, MICE, EDA/6. Feature selection with random forests.Rmd <==
#this file runs the random forest codes and stores the results into a list
#for more info on building it chec the file rf h20 prelim and results from sw rf
#this is for sliding window not expanding window
```{r}
#load("~/Thesis R/train and test data.RData")
#h2o.shutdown()
library(h2o)
h2o.init(min_mem_size='10G', max_mem_size='20G')

train = train[-c(2,3)]
htrain = as.h2o(train)
independent_var = names(train)[2:122]
dependent_var = names(train)[1]
t=127 #const
```


#trying out a two year sliding window, two years train and six months test
```{r}
get_important_features = function(total_q){
win =t*total_q #training window size
print(win)
swfs = list()

hyperparams = list(mtries= c(10, 25, 50, 100),
          max_depth = c(20, 50, 100), #max depth of tree
          ntrees = c(50, 100, 150, 200)) #number of trees

model = h2o.grid(
    algorithm = "randomForest",
    x = independent_var,
    y = dependent_var, 
    training_frame = htrain[1:win,],
    #validation_frame= htrain[(x+win):(x-1+win+shift),],
    hyper_params = hyperparams,
    seed = 786,
    nfolds= 0)

swfs[1] = model

bestswfs = h2o.getModel(swfs[[1]]@model_ids[[1]])
varimp = (bestswfs@model[["variable_importances"]])

return(varimp)
}

```

```{r}
#there are 7 different dataset lengths based on the experiments
# 36, 34, 32, 30, 28, 24, 20
varimp_36 = get_important_features(36)
varimp_34 = get_important_features(34)
varimp_32 = get_important_features(32)
varimp_30 = get_important_features(30)
varimp_28 = get_important_features(28)
varimp_24 = get_important_features(24)
varimp_20 = get_important_features(20)

save.image(file='Variable importances.RData')
```

```{r}
library(writexl)
write_xlsx(varimp_36, "~/Documents/Thesis/varimp_36.xlsx")
write_xlsx(varimp_34, "~/Documents/Thesis/varimp_34.xlsx")
write_xlsx(varimp_32, "~/Documents/Thesis/varimp_32.xlsx")
write_xlsx(varimp_30, "~/Documents/Thesis/varimp_30.xlsx")
write_xlsx(varimp_28, "~/Documents/Thesis/varimp_28.xlsx")
write_xlsx(varimp_24, "~/Documents/Thesis/varimp_24.xlsx")
write_xlsx(varimp_20, "~/Documents/Thesis/varimp_20.xlsx")

```

```{r}

write.csv(c("CR", "Ticker", "Time", varimp_36$variable),"~/Documents/Thesis/varimp_36.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_34$variable),"~/Documents/Thesis/varimp_34.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_32$variable),"~/Documents/Thesis/varimp_32.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_30$variable),"~/Documents/Thesis/varimp_30.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_28$variable),"~/Documents/Thesis/varimp_28.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_24$variable),"~/Documents/Thesis/varimp_24.csv", row.names = FALSE)

write.csv(c("CR", "Ticker", "Time", varimp_20$variable),"~/Documents/Thesis/varimp_20.csv", row.names = FALSE)

```
